docker/requirements.txt
Django==3.2
djangorestframework==3.12
djangorestframework-jwt==1.11
mysqlclient==1.4.6
Pillow 